requirements.txt
# requirements.txt

altgraph==0.17.4
anyio==4.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.47.0
gevent==23.9.1
greenlet==3.0.0rc3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
influxdb-client==1.48.0
installer==0.7.0
kiwisolver==1.4.5
loguru==0.7.3
matplotlib==3.8.2
mysql-connector-python==9.0.0
normalizer==0.2.1
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
pefile==2023.2.7
Pillow==10.0.0
py2exe==0.13.0.1
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.21
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.8
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
reactivex==4.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
speedtest-cli==2.1.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.3
websockets==11.0.3
win32_setctime==1.2.0
wxPython==4.2.1
zope.event==5.0
zope.interface==6.0
